requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
anthropic==0.7.7
openpyxl==3.1.2
xlrd==2.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0